+++ test/libsolidity/syntaxTests/visibility/external_library_function.sol
@@ -1,0 +1,14 @@
+library L {
+    function f(uint) pure external {}
+}
+
+contract C {
+    using L for *;
+
+    function f() public pure {
+        L.f(2);
+        uint x;
+        x.f();
+    }
+}
+// ----